split default tests
