+++ packages/web3-eth/test/integration/block/rpc.getTransactionCount.test.ts
@@ -15,24 +15,19 @@ You should have received a copy of the GNU Lesser General Public License
 along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { FMT_BYTES, FMT_NUMBER } from 'web3-utils';
-import { TransactionInfo, TransactionReceipt, Transaction } from 'web3-types';
+import { TransactionReceipt } from 'web3-types';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { Contract } from 'web3-eth-contract';
-import { validator } from 'web3-validator';
-
-import { Web3Eth } from '../../src';
-
+import { Web3Eth } from '../../../src';
 import {
-	getSystemTestBackend,
 	getSystemTestProvider,
 	createNewAccount,
 	createTempAccount,
 	closeOpenConnection,
-} from '../fixtures/system_test_utils';
-import { BasicAbi, BasicBytecode } from '../shared_fixtures/build/Basic';
-import { toAllVariants } from '../shared_fixtures/utils';
-import { sendFewTxes, validateTransaction } from './helper';
-import { blockSchema } from '../../src/schemas';
+} from '../../fixtures/system_test_utils';
+import { BasicAbi, BasicBytecode } from '../../shared_fixtures/build/Basic';
+import { toAllVariants } from '../../shared_fixtures/utils';
+import { sendFewTxes } from '../helper';
 
 describe('rpc with block', () => {
 	let web3Eth: Web3Eth;
@@ -101,60 +96,6 @@ describe('rpc with block', () => {
 	});
 
 	describe('methods', () => {
-		beforeAll(async () => {
-			tempAcc = await createTempAccount();
-			tempAcc2 = await createTempAccount();
-			sendOptions = { from: tempAcc.address, gas: '1000000' };
-
-			await contract.deploy(deployOptions).send(sendOptions);
-			const [receipt]: TransactionReceipt[] = await sendFewTxes({
-				web3Eth,
-				from: tempAcc.address,
-				to: tempAcc2.address,
-				value: '0x1',
-				times: 1,
-			});
-			blockData = {
-				pending: 'pending',
-				latest: 'latest',
-				earliest: 'earliest',
-				blockNumber: Number(receipt.blockNumber),
-				blockHash: String(receipt.blockHash),
-				transactionHash: String(receipt.transactionHash),
-				transactionIndex: Number(receipt.transactionIndex),
-			};
-		});
-
-		it.each(
-			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockHash' | 'blockNumber';
-				hydrated: boolean;
-				format: string;
-			}>({
-				block: ['earliest', 'latest', 'blockHash', 'blockNumber'],
-				hydrated: [true, false],
-				format: Object.values(FMT_NUMBER),
-			}),
-		)('getBlock', async ({ hydrated, block, format }) => {
-			const b = {
-				...(await web3Eth.getBlock(blockData[block], hydrated, {
-					number: format as FMT_NUMBER,
-					bytes: FMT_BYTES.HEX,
-				})),
-			};
-			if (blockData[block] === 'pending') {
-				b.nonce = '0x0';
-				b.miner = '0x0000000000000000000000000000000000000000';
-				b.totalDifficulty = '0x0';
-			}
-			expect(validator.validateJSONSchema(blockSchema, b)).toBeUndefined();
-
-			if (hydrated && b.transactions?.length > 0) {
-				// eslint-disable-next-line jest/no-conditional-expect
-				expect(b.transactions).toBeInstanceOf(Array<Transaction>);
-			}
-		});
-
 		it.each(
 			toAllVariants<{
 				block: 'earliest' | 'latest' | 'pending' | 'blockHash' | 'blockNumber';
@@ -211,55 +152,6 @@ describe('rpc with block', () => {
 			expect(Number(countAfter) - Number(countBefore)).toBe(
 				blockData[block] === 'earliest' ? 0 : count,
 			);
-		});
-
-		it.each(
-			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockHash' | 'blockNumber';
-			}>({
-				block: ['earliest', 'latest', 'pending', 'blockHash', 'blockNumber'],
-			}),
-		)('getBlockTransactionCount', async ({ block }) => {
-			const res = await web3Eth.getBlockTransactionCount(blockData[block]);
-			let shouldBe: number;
-			if (getSystemTestBackend() === 'ganache') {
-				shouldBe = blockData[block] === 'earliest' ? 0 : 1;
-			} else {
-				shouldBe = ['earliest', 'pending'].includes(String(blockData[block])) ? 0 : 1;
-			}
-			expect(Number(res)).toBe(shouldBe);
-		});
-
-		it.each(
-			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockHash' | 'blockNumber';
-			}>({
-				block: ['earliest', 'latest', 'pending', 'blockHash', 'blockNumber'],
-			}),
-		)('getBlockUncleCount', async ({ block }) => {
-			const res = await web3Eth.getBlockUncleCount(blockData[block]);
-			expect(Number(res)).toBe(0);
-		});
-
-		it.each(
-			toAllVariants<{
-				block: 'earliest' | 'latest' | 'pending' | 'blockNumber';
-			}>({
-				block: ['earliest', 'latest', 'pending', 'blockNumber'],
-			}),
-		)('getUncle', async ({ block }) => {
-			const res = await web3Eth.getUncle(blockData[block], 0);
-			// eslint-disable-next-line jest/no-standalone-expect
-			expect(res).toBeNull();
-		});
-		it('getTransactionFromBlock', async () => {
-			// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-			const tx = (await web3Eth.getTransactionFromBlock(
-				blockData.blockNumber,
-				blockData.transactionIndex,
-			))!;
-			validateTransaction(tx as TransactionInfo);
-			expect(tx?.hash).toBe(blockData.transactionHash);
 		});
 	});
 });
