+++ packages/web3-eth/test/integration/defaults.test.ts
@@ -19,7 +19,6 @@ import { Contract } from 'web3-eth-contract';
 import { hexToNumber, numberToHex, DEFAULT_RETURN_FORMAT } from 'web3-utils';
 import { TransactionBuilder, TransactionTypeParser, Web3Context, Web3PromiEvent } from 'web3-core';
 import { Hardfork, TransactionReceipt, ValidChains, Web3BaseProvider } from 'web3-types';
-import { TransactionBlockTimeoutError } from 'web3-errors';
 import {
 	detectTransactionType,
 	prepareTransactionForSigning,
@@ -33,9 +32,9 @@ import {
 	createNewAccount,
 	createTempAccount,
 	getSystemTestProvider,
-	isSocket,
-	itIf,
-	waitForOpenConnection,
+	// isSocket,
+	// itIf,
+	// waitForOpenConnection,
 } from '../fixtures/system_test_utils';
 
 import {
@@ -48,8 +47,6 @@ import { MsgSenderAbi, MsgSenderBytecode } from '../shared_fixtures/build/MsgSen
 import { getTransactionGasPricing } from '../../src/utils/get_transaction_gas_pricing';
 import { Resolve, sendFewTxes } from './helper';
 
-const MAX_32_SIGNED_INTEGER = 2147483647;
-
 describe('defaults', () => {
 	let web3Eth: Web3Eth;
 	let eth2: Web3Eth;
@@ -59,16 +56,9 @@ describe('defaults', () => {
 	let sendOptions: Record<string, unknown>;
 	let tempAcc: { address: string; privateKey: string };
 
-	beforeEach(() => {
+	beforeEach(async () => {
 		clientUrl = getSystemTestProvider();
 		web3Eth = new Web3Eth(clientUrl);
-	});
-
-	afterEach(async () => {
-		await closeOpenConnection(web3Eth);
-		await closeOpenConnection(eth2);
-	});
-	beforeEach(async () => {
 		tempAcc = await createTempAccount();
 		contract = new Contract(BasicAbi, web3Eth.getContextObject() as any);
 		deployOptions = {
@@ -76,6 +66,11 @@ describe('defaults', () => {
 			arguments: [10, 'string init value'],
 		};
 		sendOptions = { from: tempAcc.address, gas: '1000000' };
+	});
+
+	afterEach(async () => {
+		await closeOpenConnection(web3Eth);
+		await closeOpenConnection(eth2);
 	});
 
 	describe('defaults', () => {
@@ -152,7 +147,6 @@ describe('defaults', () => {
 				tempAcc2.address.toLowerCase(),
 			);
 		});
-
 		it('handleRevert', () => {
 			/*
             //TO DO: after handleRevert implementation https://github.com/ChainSafe/web3.js/issues/5069 add following tests in future release
@@ -279,7 +273,6 @@ describe('defaults', () => {
 			});
 			expect(eth2.transactionSendTimeout).toBe(120);
 		});
-
 		it('transactionBlockTimeout', () => {
 			// default
 			expect(web3Eth.transactionBlockTimeout).toBe(50);
@@ -319,7 +312,6 @@ describe('defaults', () => {
 			// eslint-disable-next-line jest/no-standalone-expect
 			expect(eth2.transactionConfirmationBlocks).toBe(4);
 		});
-
 		it('transactionConfirmationBlocks implementation', async () => {
 			const tempAcc2 = await createTempAccount();
 			const waitConfirmations = 1;
@@ -364,7 +356,6 @@ describe('defaults', () => {
 			await confirmationPromise;
 			await closeOpenConnection(eth);
 		});
-
 		it('transactionPollingInterval and transactionPollingTimeout', () => {
 			// default
 			expect(web3Eth.transactionPollingInterval).toBe(1000);
@@ -474,7 +465,6 @@ describe('defaults', () => {
 			});
 			expect(eth2.blockHeaderTimeout).toBe(4);
 		});
-
 		it('enableExperimentalFeatures', () => {
 			// default
 			expect(web3Eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout).toBe(
@@ -499,7 +489,6 @@ describe('defaults', () => {
 				true,
 			);
 		});
-
 		it('should fallback to polling if provider support `on` but `newBlockHeaders` does not arrive in `blockHeaderTimeout` seconds', async () => {
 			const tempAcc2 = await createTempAccount();
 
@@ -572,127 +561,6 @@ describe('defaults', () => {
 			expect(status).toBe(BigInt(1));
 			await closeOpenConnection(tempEth);
 		});
-
-		it('should fail if transaction was not mined within `transactionBlockTimeout` blocks', async () => {
-			const eth = new Web3Eth(clientUrl);
-			const tempAcc1 = await createTempAccount();
-			const tempAcc2 = await createTempAccount();
-
-			// Make the test run faster by casing the polling to start after 2 blocks
-			eth.transactionBlockTimeout = 2;
-
-			// Increase other timeouts so only `transactionBlockTimeout` would be reached
-			eth.transactionSendTimeout = MAX_32_SIGNED_INTEGER;
-			eth.transactionPollingTimeout = MAX_32_SIGNED_INTEGER;
-			eth.blockHeaderTimeout = MAX_32_SIGNED_INTEGER / 1000;
-
-			const from = tempAcc1.address;
-			const to = tempAcc2.address;
-			const value = `0x0`;
-
-			// Setting a high `nonce` when sending a transaction, to cause the RPC call to stuck at the Node
-			const sentTx: Web3PromiEvent<
-				TransactionReceipt,
-				SendTransactionEvents<typeof DEFAULT_RETURN_FORMAT>
-			> = eth.sendTransaction({
-				to,
-				value,
-				from,
-				// Give a high nonce so the transaction stuck forever.
-				// However, make this random to be able to run the test many times without receiving an error that indicate submitting the same transaction twice.
-				nonce: Number.MAX_SAFE_INTEGER - Math.floor(Math.random() * 100000000),
-			});
-
-			// Some providers (mostly used for development) will make blocks only when there are new transactions
-			// So, send 2 transactions, one after another, because in this test `transactionBlockTimeout = 2`.
-			// eslint-disable-next-line no-void
-			void sendFewTxes({
-				web3Eth: eth,
-				from: tempAcc2.address,
-				to: tempAcc1.address,
-				times: 2,
-				value: '0x1',
-			});
-
-			try {
-				await sentTx;
-				throw new Error(
-					'The test should fail if there is no exception when sending a transaction that could not be mined within transactionBlockTimeout',
-				);
-			} catch (error) {
-				// eslint-disable-next-line jest/no-conditional-expect
-				expect(error).toBeInstanceOf(TransactionBlockTimeoutError);
-				// eslint-disable-next-line jest/no-conditional-expect
-				expect((error as Error).message).toMatch(/was not mined within [0-9]+ blocks/);
-			}
-			await closeOpenConnection(eth);
-		});
-
-		// The code of this test case is identical to the pervious one except for `eth.enableExperimentalFeatures = true`
-		// 	And this test case will be removed once https://github.com/web3/web3.js/issues/5521 is implemented.
-		itIf(isSocket)(
-			'should fail if transaction was not mined within `transactionBlockTimeout` blocks - when subscription is used',
-			async () => {
-				const eth = new Web3Eth(clientUrl);
-				await waitForOpenConnection(eth);
-				// using subscription to get the new blocks and fire `TransactionBlockTimeoutError` is currently supported only
-				//	with `enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout` equal true.
-				eth.enableExperimentalFeatures.useSubscriptionWhenCheckingBlockTimeout = true;
-
-				const tempAcc1 = await createTempAccount();
-				const tempAcc2 = await createTempAccount();
-
-				// Make the test run faster by casing the polling to start after 2 blocks
-				eth.transactionBlockTimeout = 2;
-
-				// Increase other timeouts so only `transactionBlockTimeout` would be reached
-				eth.transactionSendTimeout = MAX_32_SIGNED_INTEGER;
-				eth.transactionPollingTimeout = MAX_32_SIGNED_INTEGER;
-				eth.blockHeaderTimeout = MAX_32_SIGNED_INTEGER / 1000;
-
-				const from = tempAcc1.address;
-				const to = tempAcc2.address;
-				const value = `0x0`;
-
-				// Setting a high `nonce` when sending a transaction, to cause the RPC call to stuck at the Node
-				const sentTx: Web3PromiEvent<
-					TransactionReceipt,
-					SendTransactionEvents<typeof DEFAULT_RETURN_FORMAT>
-				> = eth.sendTransaction({
-					to,
-					value,
-					from,
-					// Give a high nonce so the transaction stuck forever.
-					// However, make this random to be able to run the test many times without receiving an error that indicate submitting the same transaction twice.
-					nonce: Number.MAX_SAFE_INTEGER - Math.floor(Math.random() * 100000000),
-				});
-
-				// Some providers (mostly used for development) will make blocks only when there are new transactions
-				// So, send 2 transactions, one after another, because in this test `transactionBlockTimeout = 2`.
-				// eslint-disable-next-line no-void
-				void sendFewTxes({
-					web3Eth: eth,
-					from: tempAcc2.address,
-					to: tempAcc1.address,
-					times: 2,
-					value: '0x1',
-				});
-
-				try {
-					await sentTx;
-					throw new Error(
-						'The test should fail if there is no exception when sending a transaction that could not be mined within transactionBlockTimeout',
-					);
-				} catch (error) {
-					// eslint-disable-next-line jest/no-conditional-expect, jest/no-standalone-expect
-					expect(error).toBeInstanceOf(TransactionBlockTimeoutError);
-					// eslint-disable-next-line jest/no-conditional-expect, jest/no-standalone-expect
-					expect((error as Error).message).toMatch(/was not mined within [0-9]+ blocks/);
-				}
-				await closeOpenConnection(eth);
-			},
-		);
-
 		it('maxListenersWarningThreshold test default config', () => {
 			// default
 			expect(web3Eth.maxListenersWarningThreshold).toBe(100);
